pyproject.toml
[tool.poetry]
name = "rag-langchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
cloud-sql-python-connector = {extras = ["pg8000"], version = "^1.15.0"}
langchain-google-vertexai = "1.0.5"
langchain-community = "0.2.5"
pgvector = "^0.3.6"
httpx = "0.27.2"
python-dotenv = "^1.0.1"
fitz = "^0.0.1.dev2"
pymupdf = "^1.25.1"
groq = "^0.15.0"
python-multipart = "^0.0.20"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
